ScratchDetection/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
timm>=0.9.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
Pillow>=10.0.0
opencv-python>=4.8.0
torchinfo>=1.8.0
huggingface-hub>=0.16.0
